using gate authorization with product and cate controller

+++ laravel/app/Http/Controllers/CategoryController.php
@@ -12,6 +12,8 @@ class CategoryController extends Controller
     }
 
     public function createCategory(Request $request) {
+        abort_unless(auth()->user()->can('categories.create'), 403);
+        
         $category = Category::create($request->all());
         return $category;
     }
@@ -21,12 +23,16 @@ class CategoryController extends Controller
     }
 
     public function updateCategory(Request $request, $categoryId) {
+        abort_unless(auth()->user()->can('categories.update'), 403);
+        
         $category = Category::findOrFail($categoryId);
         $category->update($request->all());
         return $category;
     }
 
     public function deleteCategory($categoryId) {
+        abort_unless(auth()->user()->can('categories.update'), 403);
+        
         $category = Category::findOrFail($categoryId);
         $category->delete();
         return response()->json(['message' => 'Category deleted']);
