+++ laravel/app/Http/Controllers/ProductController.php
@@ -12,6 +12,8 @@ class ProductController extends Controller
     }
 
     public function createProduct(Request $request) {
+        abort_unless(auth()->user()->can('products.create'), 403);
+        
         $product = Product::create($request->all());
         return $product;
     }
@@ -21,12 +23,16 @@ class ProductController extends Controller
     }
 
     public function updateProduct(Request $request, $productId) {
+        abort_unless(auth()->user()->can('products.update'), 403);
+        
         $product = Product::findOrFail($productId);
         $product->update($request->all());
         return $product;
     }
 
     public function deleteProduct($productId) {
+        abort_unless(auth()->user()->can('products.update'), 403);
+        
         $product = Product::findOrFail($productId);
         $product->delete();
         return response()->json(['message' => 'Product deleted']);
